fix(connectors): restructuing project so connectors pass in important props
